Incorporate use of locality to force order date collection.
Added the capability to hide and not require
the entry of the order delivery date based on
the order being made nationally, internationally,
or either.  In combination, it is possible to
always display the order delivery date regardless
of it being required or optional having national
or international delivery.

Display is controlled first by need (ORDER_DELIVERY_DATE_LOCATION)
and then if not needed to be displayed by desire
to always display the date(ORDER_DELIVERY_DATE_DISPLAY_ALWAYS).