Consolidate catalog language defines
To reduce the number of files edited, but to also then declare definitions for use anywhere within ZC, added this file.

It is possible to streamline the file/load a little if the applicable page(s) were detected on which to load each of the above definitions.  That would at least reduce the generation of the defined value, but would also add the logic check(s) to each page load.