Consistent use of TABLE_HEADING_DELIVERY_DATE
In an effort to reduce the need to modify files, the language defines are being removed from the language specific file(s).  In doing so there was an inconsistent use of the defined variable TABLE_HEADING_DELIVERY_DATE.  In the checkout_shipping page, this variable is used with replaceable text to identify if the entry of the order delivery date is required or not.  In the checkout_confirmation page, that identification appears no longer necessary; however, because the same variable name is being used, the additional formatting code needs to still be replaced.  In this case (english) that replacement would be with empty quotes.  To maintain language "flexibility" the defined variable TABLE_HEADING_DELIVERY_DATE_EMPTY is being created and needs to be added to the language defines.  Not entirely sure how this works for other languages and if instead a completely different main define is needed to support appropriate grammar.