Update for language specific empty text
In consolidating language defines, the substitution for TABLE_HEADING_DELIVERY_DATE on the checkout_confirmation page needed to be addressed.  This offers a language dependent substitution for what is to be displayed in the checkout_confirmation page which is "designed" to consider the use of an empty text field.  This may not be the way things work for other languages and therefore some adaptation may be necessary for other languages.