Establish order_delivery_date field as a datetime
While the field is generally identified as a date field only, there is no reason to prevent the field from also containing a time if one wanted to expand to a different tool.  The associated code should still be able to support a datetime value.

This modifies the installation to use a datetime data type for order_delivery date and sets the default to be `0001-01-01 00:00:00` instead of `0001-01-01`.

Need to add an upgrade sql to modify from just a date field to a datetime field.
Fixes #1 and fixes #3.